Add files via upload
I built this in Feb 2020.I have used Adobe XD for building the project. I have built a prototype for the railway booking application on android mobile. Many of the contents we given prior by me and I visualized the features and how the application might look from user end.